+++ comparisons.py
@@ -1,0 +1,4 @@
+#Basic comparison function for testing
+def compareObjects(obj1, obj2):
+	return abs(obj1.value - obj2.value)
+
