--- test.py
@@ -1,2 +1,0 @@
-for i in range(0,5):
-   print i
